Add files via upload
This function implements the algorithm described by Schilling et. al. (Schilling et. al., 2000, Journal of theoretical biology, 203(3)) to calculate the Extreme Pathways of a metabolic model. This algorithm is relatively expensive and the number of Extreme Pathways increases drastically with network size. Therefore, this function is meant to be used only with relatively small networks.