Add files via upload
This function allows the user to calculate Fundamental Pathways as described in Schultz and Qutub, 2015, BMC systems biology, 9(1), 18. These are similar to Extreme Pathways, but are calculated in a model with no currency metabolites. With that, fewer pathways are obtained and a currency metabolite imbalance is associated with each pathway, which is also returned by this function. This function first helps guide the user as to which currency metabolites to delete, then uses expa to calculate the extreme pathways.